malloc
create characters

+++ 0x0B-malloc_free/0-create_array.c
@@ -27,7 +27,7 @@ char *create_array(unsigned int size, char c)
 	}
 	j = 0;
 
-	while (j < size)
+	while (j <(int) size)
 	{
 		s[j] = c;
 		j++;
